fase4-standalone/requirements.txt
python-tds==1.17.1
pandas==2.1.1
openpyxl==3.1.2
Flask==2.3.2
python-dotenv==1.0.0